+++ aws/doc/doc_test.go
@@ -6,6 +6,14 @@ import (
 	"github.com/wallix/awless/aws/driver"
 )
 
+func TestDocForEachCommand(t *testing.T) {
+	t.Skip()
+	for name := range awsdriver.AWSTemplatesDefinitions {
+		if doc := exampleDoc(name); len(doc) == 0 {
+			t.Errorf("missing awless CLI examples for template '%s'", name)
+		}
+	}
+}
 func TestDocForEachParam(t *testing.T) {
 	for name, def := range awsdriver.AWSTemplatesDefinitions {
 		for _, param := range def.Required() {